move llm_agent -> llm_asistant and rename everything
