+++ README.md
@@ -20,7 +20,7 @@ This repository provides a comprehensive framework for integrating Large Languag
 | `llm`           | Base module with core functionality and provider framework           |
 | `llm_tool`     | Adds tool functionality that allows LLMs to perform actions in Odoo  |
 | `llm_thread`    | Chat threads and conversation management                             |
-| `llm_agent`     | Agent management for specialized AI assistants with custom tools     |
+| `llm_agent`     | Assistant management for specialized AI assistants with custom tools     |
 | `llm_openai`    | OpenAI (GPT) provider integration                                    |
 | `llm_anthropic` | Anthropic (Claude) provider integration                              |
 | `llm_ollama`    | Ollama provider for local model deployment                           |
