--- llm_agent/changelog.rst
@@ -1,10 +1,0 @@
-16.0.1.0.1 (2025-04-04)
-~~~~~~~~~~~~~~~~~~~~~~~
-
-* [ADD] Added Agent Creator agent data record using OpenAI GPT-4o model
-* [ADD] Created data directory structure for llm_agent module
-
-16.0.1.0.0 (2025-03-01)
-~~~~~~~~~~~~~~~~~~~~~~~
-
-* [ADD] Initial release of the LLM Agent module
