--- llm_agent/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_llm_agent_user,llm.agent.user,model_llm_agent,base.group_user,1,0,0,0
-access_llm_agent_manager,llm.agent.manager,model_llm_agent,llm.group_llm_manager,1,1,1,1
