--- llm_agent/static/description/index.html
@@ -1,96 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>LLM Agent Module</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
-</head>
-<body>
-<div class="container py-4">
-    <img src="banner.jpeg" alt="LLM Agent Banner" class="img-fluid mb-4" style="max-height: 300px; width: 100%; object-fit: cover;">
-
-    <h1 class="text-primary mb-4">LLM Agent Module</h1>
-
-    <div class="bg-light p-3 rounded mb-4">
-        <p class="fw-bold">This module adds agent capabilities to the LLM integration, allowing for specialized AI assistants with dedicated tools and configurations.</p>
-    </div>
-
-    <p>The LLM Agent module extends the core LLM integration by providing a framework for creating and managing specialized AI agents. Each agent can have its own configuration, tools, and system prompts, enabling more targeted AI assistance for different use cases.</p>
-
-    <h2 class="text-primary mt-5 mb-3">Key Features</h2>
-    <ul class="list-unstyled">
-        <li>✓ Create and manage specialized AI agents</li>
-        <li>✓ Assign specific models and providers to each agent</li>
-        <li>✓ Configure agent-specific tools and capabilities</li>
-        <li>✓ Easily switch between agents in the chat interface</li>
-        <li>✓ Customize system prompts for different use cases</li>
-    </ul>
-
-    <h2 class="text-primary mt-5 mb-3">Agent Management</h2>
-    <p>The module provides comprehensive agent management capabilities:</p>
-    <ul>
-        <li>Create agents with specific names and descriptions</li>
-        <li>Assign LLM providers and models to each agent</li>
-        <li>Configure agent-specific tools and capabilities</li>
-        <li>Set custom system prompts to guide agent behavior</li>
-        <li>Organize agents by categories or departments</li>
-    </ul>
-
-    <h2 class="text-primary mt-5 mb-3">Chat Integration</h2>
-    <p>Seamlessly integrate agents into the chat interface:</p>
-    <ul>
-        <li>Select agents from a dropdown in the chat header</li>
-        <li>Switch between agents during a conversation</li>
-        <li>Agent settings automatically apply to the conversation</li>
-        <li>Visual indicators show which agent is active</li>
-    </ul>
-
-    <div class="alert alert-info mt-5">
-        <strong>Note:</strong> This module requires the base LLM integration module to be installed.
-    </div>
-
-    <h2 class="text-primary mt-5 mb-3">Security</h2>
-    <p>The module implements robust security features:</p>
-    <ul>
-        <li>Role-based access control for agent management</li>
-        <li>Regular users can use agents but only managers can create/edit them</li>
-        <li>Company-specific agent configurations</li>
-        <li>Audit logging for agent usage</li>
-    </ul>
-
-    <h2 class="text-primary mt-5 mb-3">Use Cases</h2>
-    <p>Create specialized agents for different departments and functions:</p>
-    <ul>
-        <li><strong>Customer Support Agent:</strong> Trained to answer product questions and troubleshoot issues</li>
-        <li><strong>Sales Assistant:</strong> Configured to provide product recommendations and pricing information</li>
-        <li><strong>HR Assistant:</strong> Specialized in answering policy questions and onboarding information</li>
-        <li><strong>Technical Expert:</strong> Focused on providing technical documentation and code examples</li>
-    </ul>
-
-    <h2 class="text-primary mt-5 mb-3">Configuration</h2>
-    <p>After installation:</p>
-    <ol>
-        <li>Navigate to LLM > Configuration > Agents</li>
-        <li>Create new agents with specific providers and models</li>
-        <li>Configure tools and system prompts for each agent</li>
-        <li>Assign agents to specific user groups if needed</li>
-    </ol>
-
-    <div class="alert alert-info mt-5">
-        <strong>Integration:</strong> This module extends the core LLM functionality with agent capabilities that can be used across all Odoo modules.
-    </div>
-
-    <h2 class="text-primary mt-5 mb-3">Support</h2>
-    <p>For questions or support needs, please contact:</p>
-    <p><a href="https://github.com/apexive/odoo-llm" class="text-decoration-none">https://github.com/apexive/odoo-llm</a></p>
-
-    <p class="mt-4">For more information about Apexive Solutions LLC and our other projects, visit:</p>
-    <p><a href="https://github.com/apexive/odoo-llm" class="text-decoration-none">https://github.com/apexive/odoo-llm</a></p>
-
-    <div class="text-muted mt-5 pt-3 border-top">
-        <p> 2025 Apexive Solutions LLC</p>
-    </div>
-</div>
-</body>
-</html>
