--- llm_agent/static/src/components/llm_chat_thread_header/llm_chat_thread_header.js
@@ -1,34 +1,0 @@
-/** @odoo-module **/
-
-import { LLMChatThreadHeader } from "@llm_thread/components/llm_chat_thread_header/llm_chat_thread_header";
-import { patch } from "@web/core/utils/patch";
-
-patch(LLMChatThreadHeader.prototype, "llm_agent.llm_agent_dropdown_patch", {
-  /**
-   * Get all available agents
-   */
-  get llmAgents() {
-    // Make sure we have a valid llmChat reference
-    if (!this.llmChat) {
-      return [];
-    }
-
-    // Return the agents array
-    return this.llmChat.llmAgents || [];
-  },
-
-  /**
-   * Handle agent selection
-   * @param {Object} agent - The selected agent
-   */
-  onSelectAgent(agent) {
-    this.llmChatThreadHeaderView.saveSelectedAgent(agent.id);
-  },
-
-  /**
-   * Clear the selected agent
-   */
-  onClearAgent() {
-    this.llmChatThreadHeaderView.saveSelectedAgent(false);
-  },
-});
