--- llm_agent/static/src/models/main.js
@@ -1,7 +1,0 @@
-/** @odoo-module **/
-
-// Import all models to ensure they are registered
-import "@llm_agent/models/llm_chat";
-import "@llm_agent/models/llm_agent";
-import "@llm_agent/models/thread";
-import "@llm_agent/models/llm_chat_thread_header_view";
