--- llm_agent/views/llm_menu_views.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <!-- Add Agent menu under LLM Configuration -->
-    <menuitem
-    id="menu_llm_agent"
-    name="Agents"
-    parent="llm.menu_llm_config"
-    action="action_llm_agent"
-    sequence="30"
-  />
-</odoo>
