+++ llm_assistant/changelog.rst
@@ -1,0 +1,10 @@
+16.0.1.0.1 (2025-04-04)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* [ADD] Added Assistant Creator assistant data record using OpenAI GPT-4o model
+* [ADD] Created data directory structure for llm_assistant module
+
+16.0.1.0.0 (2025-03-01)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* [ADD] Initial release of the LLM Assistant module
