+++ llm_assistant/models/__init__.py
@@ -1,0 +1,2 @@
+from . import llm_assistant
+from . import llm_thread
