+++ llm_assistant/models/llm_thread.py
@@ -1,0 +1,85 @@
+from odoo import api, fields, models
+
+
+class LLMThread(models.Model):
+    _inherit = "llm.thread"
+
+    assistant_id = fields.Many2one(
+        "llm.assistant",
+        string="Assistant",
+        ondelete="restrict",
+        help="The assistant used for this thread",
+    )
+
+    @api.onchange("assistant_id")
+    def _onchange_assistant_id(self):
+        """Update provider, model and tools when assistant changes"""
+        if self.assistant_id:
+            self.provider_id = self.assistant_id.provider_id
+            self.model_id = self.assistant_id.model_id
+            self.tool_ids = self.assistant_id.tool_ids
+
+    def set_assistant(self, assistant_id):
+        """Set the assistant for this thread and update related fields
+
+        Args:
+            assistant_id (int): The ID of the assistant to set
+
+        Returns:
+            bool: True if successful, False otherwise
+        """
+        self.ensure_one()
+
+        # If assistant_id is False or 0, just clear the assistant
+        if not assistant_id:
+            return self.write({"assistant_id": False})
+
+        # Get the assistant record
+        assistant = self.env["llm.assistant"].browse(assistant_id)
+        if not assistant.exists():
+            return False
+
+        # Update the thread with the assistant and related fields
+        update_vals = {
+            "assistant_id": assistant_id,
+            "tool_ids": [(6, 0, assistant.tool_ids.ids)],
+        }
+        if assistant.provider_id.id:
+            update_vals["provider_id"] = assistant.provider_id.id
+        if assistant.model_id.id:
+            update_vals["model_id"] = assistant.model_id.id
+        return self.write(update_vals)
+
+    def action_open_thread(self):
+        """Open the thread in the chat client interface
+
+        Returns:
+            dict: Action to open the thread in the chat client
+        """
+        self.ensure_one()
+        return {
+            "type": "ir.actions.client",
+            "tag": "llm_thread.chat_client_action",
+            "params": {
+                "default_active_id": self.id,
+            },
+            "context": {
+                "active_id": self.id,
+            },
+            "target": "current",
+        }
+    # override to include assistant's system prompt
+    def _get_system_prompt(self):
+        """Hook: return a system prompt for chat. Override in other modules. If needed"""
+        self.ensure_one()
+        system_prompt = super()._get_system_prompt()
+        
+        if self.assistant_id:
+            assistant_system_prompt = self.assistant_id.get_formatted_system_prompt()
+        
+        if assistant_system_prompt and system_prompt:
+            system_prompt = f"{assistant_system_prompt}\n\n{system_prompt}"
+        elif assistant_system_prompt:
+            system_prompt = assistant_system_prompt
+        
+        return system_prompt
