+++ llm_assistant/static/description/icon.png
