+++ llm_assistant/static/description/index.html
@@ -1,0 +1,96 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>LLM Assistant Module</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
+</head>
+<body>
+<div class="container py-4">
+    <img src="banner.jpeg" alt="LLM Assistant Banner" class="img-fluid mb-4" style="max-height: 300px; width: 100%; object-fit: cover;">
+
+    <h1 class="text-primary mb-4">LLM Assistant Module</h1>
+
+    <div class="bg-light p-3 rounded mb-4">
+        <p class="fw-bold">This module adds assistant capabilities to the LLM integration, allowing for specialized AI assistants with dedicated tools and configurations.</p>
+    </div>
+
+    <p>The LLM Assistant module extends the core LLM integration by providing a framework for creating and managing specialized AI assistants. Each assistant can have its own configuration, tools, and system prompts, enabling more targeted AI assistance for different use cases.</p>
+
+    <h2 class="text-primary mt-5 mb-3">Key Features</h2>
+    <ul class="list-unstyled">
+        <li>✓ Create and manage specialized AI assistants</li>
+        <li>✓ Assign specific models and providers to each assistant</li>
+        <li>✓ Configure assistant-specific tools and capabilities</li>
+        <li>✓ Easily switch between assistants in the chat interface</li>
+        <li>✓ Customize system prompts for different use cases</li>
+    </ul>
+
+    <h2 class="text-primary mt-5 mb-3">Assistant Management</h2>
+    <p>The module provides comprehensive assistant management capabilities:</p>
+    <ul>
+        <li>Create assistants with specific names and descriptions</li>
+        <li>Assign LLM providers and models to each assistant</li>
+        <li>Configure assistant-specific tools and capabilities</li>
+        <li>Set custom system prompts to guide assistant behavior</li>
+        <li>Organize assistants by categories or departments</li>
+    </ul>
+
+    <h2 class="text-primary mt-5 mb-3">Chat Integration</h2>
+    <p>Seamlessly integrate assistants into the chat interface:</p>
+    <ul>
+        <li>Select assistants from a dropdown in the chat header</li>
+        <li>Switch between assistants during a conversation</li>
+        <li>Assistant settings automatically apply to the conversation</li>
+        <li>Visual indicators show which assistant is active</li>
+    </ul>
+
+    <div class="alert alert-info mt-5">
+        <strong>Note:</strong> This module requires the base LLM integration module to be installed.
+    </div>
+
+    <h2 class="text-primary mt-5 mb-3">Security</h2>
+    <p>The module implements robust security features:</p>
+    <ul>
+        <li>Role-based access control for assistant management</li>
+        <li>Regular users can use assistants but only managers can create/edit them</li>
+        <li>Company-specific assistant configurations</li>
+        <li>Audit logging for assistant usage</li>
+    </ul>
+
+    <h2 class="text-primary mt-5 mb-3">Use Cases</h2>
+    <p>Create specialized assistants for different departments and functions:</p>
+    <ul>
+        <li><strong>Customer Support Assistant:</strong> Trained to answer product questions and troubleshoot issues</li>
+        <li><strong>Sales Assistant:</strong> Configured to provide product recommendations and pricing information</li>
+        <li><strong>HR Assistant:</strong> Specialized in answering policy questions and onboarding information</li>
+        <li><strong>Technical Expert:</strong> Focused on providing technical documentation and code examples</li>
+    </ul>
+
+    <h2 class="text-primary mt-5 mb-3">Configuration</h2>
+    <p>After installation:</p>
+    <ol>
+        <li>Navigate to LLM > Configuration > Assistants</li>
+        <li>Create new assistants with specific providers and models</li>
+        <li>Configure tools and system prompts for each assistant</li>
+        <li>Assign assistants to specific user groups if needed</li>
+    </ol>
+
+    <div class="alert alert-info mt-5">
+        <strong>Integration:</strong> This module extends the core LLM functionality with assistant capabilities that can be used across all Odoo modules.
+    </div>
+
+    <h2 class="text-primary mt-5 mb-3">Support</h2>
+    <p>For questions or support needs, please contact:</p>
+    <p><a href="https://github.com/apexive/odoo-llm" class="text-decoration-none">https://github.com/apexive/odoo-llm</a></p>
+
+    <p class="mt-4">For more information about Apexive Solutions LLC and our other projects, visit:</p>
+    <p><a href="https://github.com/apexive/odoo-llm" class="text-decoration-none">https://github.com/apexive/odoo-llm</a></p>
+
+    <div class="text-muted mt-5 pt-3 border-top">
+        <p> 2025 Apexive Solutions LLC</p>
+    </div>
+</div>
+</body>
+</html>
