+++ llm_assistant/static/src/components/llm_chat_thread_header/llm_chat_thread_header.xml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<templates xml:space="preserve">
+    <t
+    t-name="llm_assistant.LLMChatThreadHeader"
+    t-inherit="llm_thread.LLMChatThreadHeader"
+    t-inherit-mode="extension"
+  >
+        <xpath
+      expr="//div[hasclass('o_ThreadDropdowns')]/div[hasclass('dropdown')][1]"
+      position="before"
+    >
+            <!-- LLM Assistant Dropdown -->
+            <div class="dropdown me-2">
+                <button
+          class="btn btn-light dropdown-toggle d-flex align-items-center gap-2"
+          type="button"
+          data-bs-toggle="dropdown"
+          aria-expanded="false"
+          style="max-width: 200px;"
+        >
+                    <div class="text-truncate">
+                        <t t-if="llmChatThreadHeaderView.selectedAssistant">
+                            <t
+                t-esc="llmChatThreadHeaderView.selectedAssistant.name"
+              />
+                        </t>
+                        <t t-else="">
+                            Select Assistant
+                        </t>
+                    </div>
+                </button>
+                <ul class="dropdown-menu">
+                    <li>
+                        <a
+              class="dropdown-item text-truncate"
+              href="#"
+              t-on-click="onClearAssistant"
+              t-attf-class="{{ !llmChatThreadHeaderView.selectedAssistant ? 'active' : '' }}"
+            >
+                            <i class="fa fa-times me-1" /> No Assistant
+                        </a>
+                    </li>
+                    <li><hr class="dropdown-divider" /></li>
+                    <t t-foreach="llmAssistants" t-as="assistant" t-key="assistant.id">
+                        <li>
+                            <a
+                class="dropdown-item text-truncate"
+                href="#"
+                t-on-click="() => this.onSelectAssistant(assistant)"
+                t-att-title="assistant.name"
+                t-attf-class="{{ llmChatThreadHeaderView.selectedAssistant and llmChatThreadHeaderView.selectedAssistant.id === assistant.id ? 'active' : '' }}"
+              >
+                                <t t-esc="assistant.name" />
+                            </a>
+                        </li>
+                    </t>
+                </ul>
+            </div>
+        </xpath>
+    </t>
+</templates>
