+++ llm_assistant/static/src/models/main.js
@@ -1,0 +1,7 @@
+/** @odoo-module **/
+
+// Import all models to ensure they are registered
+import "@llm_assistant/models/llm_chat";
+import "@llm_assistant/models/llm_assistant";
+import "@llm_assistant/models/thread";
+import "@llm_assistant/models/llm_chat_thread_header_view";
