+++ llm_assistant/views/llm_menu_views.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <!-- Add Assistant menu under LLM Configuration -->
+    <menuitem
+    id="menu_llm_assistant"
+    name="Assistants"
+    parent="llm.menu_llm_config"
+    action="action_llm_assistant"
+    sequence="30"
+  />
+</odoo>
