+++ llm_tool/static/description/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Automate Your Odoo Database with AI Agents & Chat AI | ChatGPT, Grok, Anthropic, DeepSeek</title>
+    <title>Automate Your Odoo Database with AI Assistants & Chat AI | ChatGPT, Grok, Anthropic, DeepSeek</title>
     <meta name="description" content="Boost your Odoo database automation with AI-powered agents using ChatGPT, Grok, Anthropic, and DeepSeek. Streamline workflows, optimize data management, and enhance productivity with AI tools seamlessly integrated into your Odoo instance;">
     <meta name="keywords" content="Odoo AI automation, Odoo database AI, AI agents for Odoo, ChatGPT Odoo integration, Grok AI Odoo, Anthropic AI for Odoo, DeepSeek Odoo tools, AI-powered ERP, automate Odoo workflows, AI chat for Odoo, Odoo AI assistant, AI-driven ERP management, intelligent Odoo agents, AI for business automation, function calling, tool execution">
     <meta name="author" content="Apexive Solutions LLC">
@@ -13,7 +13,7 @@
         <!-- Header Section with Banner -->
         <header class="text-center mb-5">
             <h1 class="display-4">LLM Tool</h1>
-            <p class="lead">Automate Your Odoo Database with AI Agents & Chat AI | ChatGPT, Grok, Anthropic, DeepSeek</p>
+            <p class="lead">Automate Your Odoo Database with AI Assistants & Chat AI | ChatGPT, Grok, Anthropic, DeepSeek</p>
             <img src="banner.jpeg" alt="Easy AI Chat Interface for Odoo" class="img-fluid rounded mb-3" style="max-height: 300px; object-fit: cover;">
         </header>
 
@@ -33,7 +33,7 @@
                     <ul class="list-group">
                         <li class="list-group-item"><strong>Function Calling:</strong> Enable AI models to call specific functions based on user requests</li>
                         <li class="list-group-item"><strong>Odoo AI Automation:</strong> Define and manage LLM tools with custom implementations</li>
-                        <li class="list-group-item"><strong>AI Agents for Odoo:</strong> Dynamic schema generation from Pydantic models</li>
+                        <li class="list-group-item"><strong>AI Assistants for Odoo:</strong> Dynamic schema generation from Pydantic models</li>
                         <li class="list-group-item"><strong>Chat AI Integration:</strong> Flexible override options for tool descriptions and schemas</li>
                         <li class="list-group-item"><strong>AI-Powered ERP:</strong> Integration with Odoo mail threads for intelligent chat interactions</li>
                         <li class="list-group-item"><strong>Extensible Design:</strong> Add new tool implementations for your business needs</li>
